Some issues are fixed in the branch
Based on Sept08th build, some issues found when the test did registration oprional parameters adding. That the test could be still passed if the request message has wrong parameters name or bad values. Add some code to check the final number of parameters to be added if they are not in the test json file. If request message has bad value or wrong name, an error will be given.

But right now there is limitation for the optional parameters in the message of grantRequest and spectrumInquiryRequest. inquiredSpectrum and operationParam can not be added properly. A fix will be considered in future.