Fix for required and optional params in nested request json message (Issue 46)
Fix for Issue #46.

There needs to be some extra logic in Assert.py in order to handle a
nested parameters, where the Request mesasage contains  json nested
parameters (such as installationParam) where some  are Required (in
required Json) and some of the parameters are optional (not in required
Json, but in optional schema).

I think this should only impact REG test cases, since the other test
cases don't need to handle this case.