Fixed functionality for HBT Absent test cases
Updated code when HBT absent is triggered.   Should not impact other
test cases.

Note:  works for 1 CBSD, but not for more than 1.  Since v1.1 test cases
only call for 1 CBSD for HBT absent, this is a partial fix.  Should be
revisited as an enhancement in future to fix multiple-CBSD test case