Adding Test Harness Version number & date to log files and console (Issue #123)
Added version number and date to Consts.py, and print to console and log
files.

Note:  need to manually update Consts.py for every release.  Use test
"develop" prior to version number to indicate a working version (not an
official release).

Printed in startTest() of DebugLogger.py and CmdLogger.py.